requirements.txt
deepface==0.0.91
dlib==19.22.0
firebase_admin==6.5.0
numpy==1.24.3
opencv_contrib_python==4.10.0.84
opencv_python==4.9.0.80
Pillow==10.4.0
PyJWT==2.8.0
Pyrebase4==4.8.0
scipy==1.10.0
